Rename dp_models package to data_platform
Since the package contains more than just the data platform models we
should rename it so that the name is not misleading.
