+++ src/ggrc/data_platform/attribute_templates.py
@@ -6,7 +6,7 @@
 from sqlalchemy import orm
 
 from ggrc import db
-from ggrc.dp_models import base
+from ggrc.data_platform import base
 
 
 class AttributeTemplates(base.Base, db.Model):
