+++ src/ggrc/data_platform/attributes.py
@@ -7,7 +7,7 @@ from sqlalchemy import orm
 from sqlalchemy.ext.declarative import declared_attr
 
 from ggrc import db
-from ggrc.dp_models import base
+from ggrc.data_platform import base
 
 
 class Attributes(base.Base, db.Model):
