+++ src/ggrc/models/all_models.py
@@ -7,13 +7,13 @@ import sys
 
 from ggrc.access_control.role import AccessControlRole
 from ggrc.access_control.list import AccessControlList
-from ggrc.dp_models.attribute_definitions import AttributeDefinitions
-from ggrc.dp_models.attribute_templates import AttributeTemplates
-from ggrc.dp_models.attribute_types import AttributeTypes
-from ggrc.dp_models.attributes import Attributes
-from ggrc.dp_models.namespaces import Namespaces
-from ggrc.dp_models.object_templates import ObjectTemplates
-from ggrc.dp_models.object_types import ObjectTypes
+from ggrc.data_platform.attribute_definitions import AttributeDefinitions
+from ggrc.data_platform.attribute_templates import AttributeTemplates
+from ggrc.data_platform.attribute_types import AttributeTypes
+from ggrc.data_platform.attributes import Attributes
+from ggrc.data_platform.namespaces import Namespaces
+from ggrc.data_platform.object_templates import ObjectTemplates
+from ggrc.data_platform.object_types import ObjectTypes
 from ggrc.models import inflector
 from ggrc.models.access_group import AccessGroup
 from ggrc.models.assessment import Assessment
